### 2017-08-20 #145 @daranzolin
rperseus -- Get texts from the Perseus Digital Library
-   `rperseus` extracts classical texts from the [Perseus Digital Library's CapiTainS environment](http://www.perseus.tufts.edu/hopper/) and returns them as a data frame.

```
Package: rperseus
Title: Gets texts from the Perseus Digital Library
Version: 0.0.0.9000
Authors@R: person("David", "Ranzolin", email = "daranzolin@gmail.com", role = c("aut", "cre"))
Description: Perseus is a collection of primary source texts. This package helps
    you get them.
    httr,
    xml2,
    rvest,
    tibble
Suggests: knitr,
    covr
```
https://github.com/daranzolin/rperseus

This package falls under **data retrieval.** With the recent surge in text mining and sentiment analysis, the works of past authors are now at our disposal.

Classicists with technical training, as well as enthusiasts interested in text mining and analysis. The data obtained from this package could serve as a springboard to linguistic and literary analysis.

None that I'm aware of.

- [x ] does not violate the Terms of Service of any service it interacts with.
- [x ] has a CRAN and OSI accepted license.
- [x ] contains a README with instructions for installing the development version.
- [x ] includes documentation with examples for all functions.
- [x ] contains a vignette with examples of its essential functions and uses.
- [x ] has a test suite.
- [x ] has continuous integration, including reporting of test coverage, using services such as Travis CI, Coeveralls and/or CodeCov.
- [x ] I agree to abide by [ROpenSci's Code of Conduct](https://github.com/ropensci/onboarding/blob/master/policies.md#code-of-conduct) during the review process and in maintaining my package should it be accepted.

- [x ] Do you intend for this package to go on CRAN?

- [x ] Does `R CMD check` (or `devtools::check()`) succeed?  Paste and describe any errors or warnings:

- [x ] Does the package conform to [rOpenSci packaging guidelines](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md)? Please describe any exceptions:
