### 2017-08-24 #145 @noamross
### Editor checks:

---

#### Editor comments

Thank you for the submission @daranzolin! This package is a good fit and I am currently seeking reviewers.

Some notes: Below is output from `goodpractice::gp()`.  It shows a generally clean output.  A few things to be addressed, though:
-   running lintr::lint_package() shows some additional extraneous whitespace and commented-out code
-  You may want to use linter exclusions (See the linter README: https://github.com/jimhester/lintr) for lines with long URLs
-  The "Namespaces in Imports" field error should be addressed. It appears you import packages that you do not use directly.  This should be fixed before I assign reviewers.
- I believe the UTF NOTE is handled correctly and this is just a snafu in **goodpractice**, but I will seek a reviewer with related expertise in ensuring that all unicode is handled properly.

```
── GP rperseus ────────────────────────────────────

It is good practice to

    R/utils.R:28:NA
    R/utils.R:39:NA

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide.
    Try make your lines shorter than 80 characters

    R/get_perseus_text.R:17:1
    R/get_perseus_text.R:22:1
    R/utils.R:37:1
    R/utils.R:45:1
    tests/testthat/test-get-perseus-text.R:17:1
    ... and 1 more lines

  ✖ not import packages as a whole, as this can cause name clashes between the imported packages. Instead, import only the specific
    functions you need.
  ✖ fix this R CMD check NOTE: Namespaces in Imports field not imported from: ‘stats’ ‘tibble’ ‘tidyr’ All declared Imports should
    be used.
  ✖ fix this R CMD check NOTE: Note: found 61 marked UTF-8 strings
───────────────────────────────────────────────
```
---

Reviewers:
Due date:
