### 2017-08-25 7df1bd7
removed comented out code
