### 2017-08-25 a6fe285
cleaned code for lintr
