+++ R/get_perseus_text.R
@@ -14,12 +14,20 @@
 #' }
 get_perseus_text <- function(text_urn) {
 
-  if (!text_urn %in% internal_perseus_catalog$urn) stop("invalid text_urn argument: check perseus_catalog for valid URNs")
+  if (!text_urn %in% internal_perseus_catalog$urn) {
+    stop("invalid text_urn argument: check perseus_catalog for valid URNs")
+  }
 
   new_urn <- reformat_urn(text_urn)
   text_index <- get_full_text_index(new_urn)
   if (grepl("NA", text_index)) stop("No text available.")
-  text_url <- sprintf("http://cts.perseids.org/api/cts/?request=GetPassage&urn=%s:%s", text_urn, text_index)
+  BASE_URL <- "http://cts.perseids.org/api/cts"
+  text_url <- httr::modify_url(BASE_URL,
+                   query = list(
+                     request = "GetPassage",
+                     urn = paste(text_urn, text_index, sep = ":")
+                     )
+                   )
   text_df <- extract_text(text_url) %>%
     dplyr::mutate(urn = text_urn) %>%
     dplyr::left_join(internal_perseus_catalog, by = "urn") %>%
