### 2017-08-25 #145 @daranzolin
Much obliged @noamross! I updated the package accordingly: `lintr::lint_package` returns nothing, and I removed the extraneous imports.
