### 2017-08-31 #145 @czeildi
Overall I think this is a great package as it is very focused: does one thing and does it well. I could immediately get going based on the readme. I found it great and helpful that the package also has a website generated by pkgdown. I have a few suggestions regarding style, please see my comments at the end and after the packaging guidelines section.

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

>- The package should contain top-level documentation for ?foobar, (or ?foobar-package if there is a naming conflict). Optionally, you can use both ?foobar and ?foobar-package for the package level manual file, using @aliases roxygen tag.
>- Make sure to list packages used for testing (testthat), and documentation (knitr, **roxygen2**) in your Suggests section of package dependencies.
>- Make sure you include links to websites if you wrap a web API, scrape data from a site, etc. in the Description field of your DESCRIPTION file
>- The readme should start with the package name

---

These are my suggestions and things to consider in order of importance:

* consider adding a different error message when the text_urn is valid but the API somehow is not available
* use `stringr::str_replace_all` instead of gsub for consistency with other used stringr functions and more compatibility with the `%>%`.
* data documentation:
    * document the columns in the order they appear
    * one / missing from the documentation of language
    * `...` is not needed as all columns are explicitly documented
    * based on [this](http://r-pkgs.had.co.nz/data.html), consider adding your perseus\_catalog [generator script](https://github.com/daranzolin/rperseus_catalog/blob/master/perseus_catalog.R) under `raw_data` so that the code for this package is fully self-contained
* consider extracting the following part to a util function so that each function is more focused:
```r
BASE_URL <- "http://cts.perseids.org/api/cts"
  text_url <- httr::modify_url(BASE_URL,
                   query = list(
                     request = "GetPassage",
                     urn = paste(text_urn, text_index, sep = ":")
                     )
                   )
```

```r
get_text_url <- function(text_urn, text_index) {
    # ...
}
```
This extraction could make the following test more clear as well: "Lots Jesus talk extracted from the Gospel of Matthew"

* do not use explicit return statements: they are usually saved for early return from functions
* extend the documentation of `get_perseus_text` in what does it return: something like whole text in sections with metadata in columns
* small typo in the description part of readme: API endpoint instead of API end point
* consider adding internal roxygen2 documentation to `get_full_text_index` as it was the hardest for me to understand even after running it for multiple examples step-by-step.
* news should contain information most relevant to the user so I would not include that unit test have been added
* consider group\_name instead of groupname as a column name in your perseus\_catalog data
* consider renaming your local 'r' variable to response or similar
* I also have a question: row\_number in the last but one row of `get_perseus_text`: which package does it come from as I believe it is not the row\_number function from dplyr? Maybe consider using that for consistency.
* further improvement idea, maybe for a later version: make it possible to directly query by label (and language) instead of urn
