+++ R/data.R
@@ -4,12 +4,11 @@
 #'
 #' @format A data frame with 2291 rows and 5 variables:
 #' \describe{
+#'   \item{urn}{Uniform Resource Number}
 #'   \item{groupname}{Could refer to author (e.g. "Aristotle") or corpus (e.g. "New Testament")}
 #'   \item{label}{Text label, e.g. "Phaedrus"}
-#'   \item{urn}{Uniform Resource Number}
-#'   \item{language}{language, "lat"(Latin)/"grc"(Greek)"heb"(Hebrew)}
 #'   \item{description}{Text description}
-#'   ...
+#'   \item{language}{Text language, e.g. "grc" = Greek, "lat" = Latin, "eng" = English}
 #' }
 #' @source \url{http://cts.perseids.org/api/cts/?request=GetCapabilities}
 "perseus_catalog"
