### 2017-09-01 ad412f8
updated NEWS
