+++ NEWS.md
@@ -4,5 +4,4 @@
   * `get_perseus_text` function, for getting texts from the Perseus Digital Library
   * `perseus_catalog` data frame to peruse text URNs
   * Introductory vignette including basic examples of retrieving texts
-  * Unit tests for `get_perseus_text`
 * Added a `NEWS.md` file to track changes to the package.
