### 2017-09-01 8d1548d
simplified tests
