+++ tests/testthat/test-utils.R
@@ -7,18 +7,10 @@ test_that("Correct index is scraped and returned by get_full_text_index", {
   expect_equal(ind, "172-223")
 })
 
-test_that("Lots Jesus talk extracted from the Gospel of Matthew", {
+test_that("Lots of Jesus talk extracted from the Gospel of Matthew", {
   skip_on_cran()
 
-  BASE_URL <- "http://cts.perseids.org/api/cts/"
-  text_url <- httr::modify_url(
-    BASE_URL,
-    query = list(
-      request = "GetPassage",
-      urn = "urn:cts:greekLit:tlg0031.tlg001.perseus-eng2:1.1-28.20"
-    )
-  )
-
+  text_url <- get_text_url("urn:cts:greekLit:tlg0031.tlg001.perseus-eng2", "1.1-28.20")
   jc <- extract_text(text_url)
   jc <- length(purrr::flatten_chr(stringr::str_extract_all(jc$text, "Jesus")))
   expect_gt(jc, 100)
