### 2017-09-01 4f7db24
added roxygen template, new function get_text_url
