### 2017-09-01 d73e3bb
updated vignette
