+++ R/get_perseus_text.R
@@ -2,7 +2,16 @@
 #'
 #' @param text_urn Valid uniform resource number (URN) obtained from perseus_catalog.
 #'
-#' @return a seven column tbl_df
+#' @return A seven column tbl_df with one row for each "section" (splits vary from text--could be line, chapter, etc.).
+#' Columns:
+#' \describe{
+#'   \item{text}{character vector of text}
+#'   \item{urn}{Uniform Resource Number}
+#'   \item{group_name}{Could refer to author (e.g. "Aristotle") or corpus (e.g. "New Testament")}
+#'   \item{label}{Text label, e.g. "Phaedrus"}
+#'   \item{description}{Text description}
+#'   \item{language}{Text language, e.g. "grc" = Greek, "lat" = Latin, "eng" = English}
+#' }
 #' @import dplyr
 #' @export
 #'
@@ -12,6 +21,7 @@
 #' get_perseus_text("urn:cts:greekLit:tlg0013.tlg028.perseus-grc2")
 #' get_perseus_text("urn:cts:latinLit:stoa0215b.stoa003.opp-lat1")
 #' }
+
 get_perseus_text <- function(text_urn) {
 
   if (!text_urn %in% internal_perseus_catalog$urn) {
@@ -21,16 +31,10 @@ get_perseus_text <- function(text_urn) {
   new_urn <- reformat_urn(text_urn)
   text_index <- get_full_text_index(new_urn)
   if (grepl("NA", text_index)) stop("No text available.")
-  BASE_URL <- "http://cts.perseids.org/api/cts"
-  text_url <- httr::modify_url(BASE_URL,
-                   query = list(
-                     request = "GetPassage",
-                     urn = paste(text_urn, text_index, sep = ":")
-                     )
-                   )
-  text_df <- extract_text(text_url) %>%
+  text_url <- get_text_url(text_urn, text_index)
+  df <- extract_text(text_url) %>%
     dplyr::mutate(urn = text_urn) %>%
     dplyr::left_join(internal_perseus_catalog, by = "urn") %>%
     dplyr::mutate(section = row_number())
-  return(text_df)
+  df
 }
