+++ man/perseus_catalog.Rd
@@ -6,12 +6,11 @@
 \title{Metadata for texts available via the Perseus Digital Library.}
 \format{A data frame with 2291 rows and 5 variables:
 \describe{
-  \item{groupname}{Could refer to author (e.g. "Aristotle") or corpus (e.g. "New Testament")}
-  \item{label}{Text label, e.g. "Phaedrus"}
   \item{urn}{Uniform Resource Number}
-  \item{language}{language, "lat"(Latin)/"grc"(Greek)"heb"(Hebrew)}
+  \item{group_name}{Could refer to author (e.g. "Aristotle") or corpus (e.g. "New Testament")}
+  \item{label}{Text label, e.g. "Phaedrus"}
   \item{description}{Text description}
-  ...
+  \item{language}{Text language, e.g. "grc" = Greek, "lat" = Latin, "eng" = English}
 }}
 \source{
 \url{http://cts.perseids.org/api/cts/?request=GetCapabilities}
