### 2017-09-01 #145 @daranzolin
My sincere thanks to Ildikó, both for the speed and comprehensiveness of the review. I've learned a lot about best practices already!

I've tried to address each concern in the latest commits:

> use stringr::str_replace_all instead of gsub for consistency with other used stringr functions and more compatibility with the %>%.

Done.

> data documentation:
document the columns in the order they appear
one / missing from the documentation of language
... is not needed as all columns are explicitly documented
based on this, consider adding your perseus_catalog generator script under raw_data so that the code for this package is fully self-contained

Done.

> consider extracting...to a util function so that each function is more focused (get_text_url)

Done.

> do not use explicit return statements: they are usually saved for early return from functions

Done. Is it best practice to leave the object at the end? Or the last function that produced it?

> extend the documentation of get_perseus_text in what does it return: something like whole text in sections with metadata in columns

Done.

> small typo in the description part of readme: API endpoint instead of API end point

Done.

> consider adding internal roxygen2 documentation to get_full_text_index as it was the hardest for me to understand even after running it for multiple examples step-by-step.

Done. Is that also best practice? What criteria should I use to document unexported functions?

> news should contain information most relevant to the user so I would not include that unit test have been added

Done.

> consider group_name instead of groupname as a column name in your perseus_catalog data

Done.

> consider renaming your local 'r' variable to response or similar

Done.

> I also have a question: row_number in the last but one row of get_perseus_text: which package does it come from as I believe it is not the row_number function from dplyr? Maybe consider using that for consistency.

I admit I am also confused by this. I assumed it *was* from dplyr, but when I substituted `dplyr::row_number`, the function throws an error:
```
Error in mutate_impl(.data, dots) :
  Evaluation error: argument "x" is missing, with no default.
```
A reprex:

```
df <- data.frame(x = runif(10))
df %>% dplyr::mutate(rn = row_number())
           x rn
1  0.4389244  1
2  0.1549839  2
3  0.8573690  3
4  0.7328806  4
5  0.9063636  5
6  0.3541390  6
7  0.5007385  7
8  0.3527843  8
9  0.7526521  9
10 0.1082413 10

```
Now explicitly indicating the namespace:

```

df <- data.frame(x = runif(10))
df %>% dplyr::mutate(rn = dplyr::row_number())
Error in mutate_impl(.data, dots) :
  Evaluation error: argument "x" is missing, with no default.

```

I'm not sure what exactly is happening here.

> further improvement idea, maybe for a later version: make it possible to directly query by label (and language) instead of urn

This crossed my mind too, but my initial thoughts were that trying to figure out the precise spelling
of an author and text would be less user-friendly than just scanning for and pasting the urn.
