### 2017-09-01 bddcd16
included httr::stop_for_status
