### 2017-09-01 #145 @czeildi
hi, just a quick comment about `row_number`: `dplyr::row_number` should be called with one of the columns of your data frame, e.g. `section = dplyr::row_number(urn)`. My question still holds but your solution seems perfectly fine, I just do not understand this. I will look at the other parts of your update later.
