### 2017-09-02 #145 @czeildi
Hi @daranzolin , the changes look great!

You forgot to apply the groupname --> group_name change in the readme and vignette as well so these now do not run.

Answering your question about return statements: I believe it is accepted practice not to use explicit return although I failed to quickly find reference for this. From an R function the last evaluated expression will be returned so for example the last lines of `extract_text` could be this:

```r
dplyr::data_frame(text = text) %>%
    dplyr::filter(text != "")
```
Usually I name my return variables only if the name can add meaningful information or if there are many piped statements after each other.

Regarding documenting internal functions: I am not aware of any guideline or best practices, in my personal experince I found it useful to document functions which are especially difficult to understand. But this is quite subjective.

I believe your code needs a few minor changes to pass packaging guidelines:

>- The package should contain top-level documentation for ?foobar, (or ?foobar-package if there is a naming conflict). Optionally, you can use both ?foobar and ?foobar-package for the package level manual file, using @aliases roxygen tag.

You can find direction [here](http://r-pkgs.had.co.nz/man.html#man-packages)

>- Make sure to list packages used for testing (testthat), and documentation (knitr, **roxygen2**) in your Suggests section of package dependencies.

I also noticed that your readme loads tidyverse and tidytext, you should list the packages you use in your readme and vignettes under Suggests field. [Reference](http://r-pkgs.had.co.nz/description.html#dependencies). I believe from tidyverse you only use dplyr an purrr which are already imported so I advise load these in your readme and not the whole tidyverse.

>- Make sure you include links to websites if you wrap a web API, scrape data from a site, etc. in the Description field of your DESCRIPTION file
>- The readme should start with the package name

I also have a couple more comments:

 - I see you did go forward with my suggestion about `stringr::str_replace_all`, in this case you might want to follow through and use it in `extract_text` as well. You use the `%>%` a lot in your code so I think it would make these lines slightly more readable:
```r
text <- purrr::map(r_list$reply$passage$TEI$text$body$div,
                     ~ paste(unlist(.), collapse = " ")) %>%
    stringr::str_replace_all("\\s+", " ") %>%
    stringr::str_replace_all("\\*", "") %>%
    stringr::str_trim()
```
 - I see that you included `dplyr::` in many places. In this case I believe you do not have to import the whole dplyr package but only pipe from dplyr: `@importFrom dplyr %>%` This avoids unnecessary name collisions when someone loads your package but do not want to use dplyr. (Although dplyr is very widely used, but generally it is good practice to have as few dependencies as possible which do not hurt convenience much.) The package level documentation file is a great place to include imports like `%>%` which you use in many of your functions.

These comments are more a question of taste so feel free to apply those you can agree with. Great work!
