+++ README.Rmd
@@ -43,7 +43,7 @@ library(tidyverse)
 library(rperseus)
 
 aeneid_latin <- perseus_catalog %>% 
-  filter(groupname == "Virgil",
+  filter(group_name == "Virgil",
          label == "Aeneid",
          language == "lat") %>% 
   pull(urn) %>% 
@@ -54,7 +54,7 @@ You can also request an English translation for some texts:
 
 ```{r eval = FALSE}
 aeneid_english <- perseus_catalog %>% 
-  filter(groupname == "Virgil",
+  filter(group_name == "Virgil",
          label == "Aeneid",
          language == "eng") %>% 
   pull(urn) %>% 
@@ -69,7 +69,7 @@ Refer to the language variable in `perseus_catalog` for translation availability
 
 ```{r eval = FALSE}
 plato <- perseus_catalog %>% 
-  filter(groupname == "Plato",
+  filter(group_name == "Plato",
          language == "eng") %>% 
   pull(urn) %>% 
   map_df(get_perseus_text)
@@ -81,7 +81,7 @@ And here's how to retrieve the Greek text from Sophocles' underrated *Philoctete
 library(tidytext)
 
 philoctetes <- perseus_catalog %>% 
-  filter(groupname == "Sophocles",
+  filter(group_name == "Sophocles",
          label == "Philoctetes",
          language == "grc") %>% 
   pull(urn) %>%
