+++ NAMESPACE
@@ -1,4 +1,4 @@
 # Generated by roxygen2: do not edit by hand
 
 export(get_perseus_text)
-import(dplyr)
+importFrom(dplyr,"%>%")
