+++ README.Rmd
@@ -39,7 +39,8 @@ devtools::install_github("daranzolin/rperseus")
 To obtain a particular text, you must first know its full Uniform Resource Name (URN). URNs can be perused in the `perseus_catalog`, a data frame lazily loaded into the package. For example, say I want a copy of Virgil's *Aeneid*:
 
 ```{r warning = FALSE, message=FALSE}
-library(tidyverse)
+library(dplyr)
+library(purrr)
 library(rperseus)
 
 aeneid_latin <- perseus_catalog %>% 
