### 2017-09-04 c52dccf
adjusted imports from dplyr, added documentation
