+++ man/perseus_catalog.Rd
@@ -1,7 +1,8 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
+% Please edit documentation in R/data.R, R/rperseus.R
 \docType{data}
 \name{perseus_catalog}
+\alias{perseus_catalog}
 \alias{perseus_catalog}
 \title{Metadata for texts available via the Perseus Digital Library.}
 \format{A data frame with 2291 rows and 5 variables:
@@ -20,5 +21,7 @@ perseus_catalog
 }
 \description{
 A dataset containing the texts available from the Perseus Digital Library.
+
+Perseus Catalog
 }
 \keyword{datasets}
