### 2017-09-04 43851b2
added package-level documentation
