+++ R/rperseus.R
@@ -1,0 +1,16 @@
+#' rperseus: A package to obtain texts from the Perseus Digital Library
+#'
+#' The package contains a catalog and a function. The catalog is \code{perseus_catalog} and can be
+#' perused to locate a particular text and its corresponding URN. Then pass the URN to \code{get_perseus_text}
+#' to obtain the text.
+#'
+#'
+#' @docType package
+#' @name rperseus
+#' @author David Ranzolin \email{daranzolin@@gmail.com}
+NULL
+#' Perseus Catalog
+#' @name perseus_catalog
+#' @docType data
+#' @keywords datasets
+NULL
