+++ man/rperseus.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rperseus.R
+\docType{package}
+\name{rperseus}
+\alias{rperseus}
+\alias{rperseus-package}
+\title{rperseus: A package to obtain texts from the Perseus Digital Library}
+\description{
+The package contains a catalog and a function. The catalog is \code{perseus_catalog} and can be
+perused to locate a particular text and its corresponding URN. Then pass the URN to \code{get_perseus_text}
+to obtain the text.
+}
+\author{
+David Ranzolin \email{daranzolin@gmail.com}
+}
