### 2017-09-04 #145 @daranzolin
Woops, thanks for the reminder! Fixed.

Regarding package level documentation--[is this correct?](https://github.com/daranzolin/rperseus/blob/master/R/rperseus.R)

Also addressed:
1. Added packages to "suggests" in DESCRIPTION
2. Included link to website in DESCRIPTION
3. README starts with package name
