### 2017-09-05 #145 @noamross
Thanks for your rapid review, @czeildi, and your rapid responses @daranzolin! Just an FYI, you may want to hold off on making additional changes in response  until the second reviewer gets his review in, as its possible there are related or conflicting opinions.
