### 2017-09-19 4c08d4d
fix R CMD check NOTE: no visible binding for urn
Explicitly adding `.data$urn` in the `mutate` call to avoid the following NOTE when running R CMD check:

```
get_perseus_text: no visible binding for global variable ‘urn’
Undefined global functions or variables:
  urn
```
