### 2017-09-19 #145 @fmichonneau
I think the goal of the package could be made a little clearer by being more direct. Something like: "`rperseus` can be used to download classic texts (and in some case their English translation) from Homer to Cicero to Boetheius, to be analyzed in R." Also define, or link to what CapiTainS is.

  * `get_full_text_index` isn't exported but documented. I think a more
    realistic example of this function (if it's actually useful) could be
    included. Indeed, it seems to be useful when it's part of a pipe (otherwise,
    how would a user construct a modified urn?) that it would be good to see
    reflected in the example.
  * in the documentation of the package, maybe start by describing the goal of
    the package before writing about what it contains. You could also consider
    pointing to the vignette to let users know how to get started.
  * I think the dataset (`perseus_catalog`) is documented twice, once in
    `R/data.R` and once in `R/rperseus.R`.

  * It seems it would be useful to show in the examples of `get_perseus_text()`
    how to provide URNs from the dataset.

  * I don't see a CONTRIBUTING file or mention on how to contribute in the
    README.

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

  * add a link to `CONDUCT.md` in the `README` file

---

This is a simple and elegant package to retrieve classical full texts. The package is straightforward to use, and the vignette makes it easy to get started. I have a few comments in the checklist above and below. I found the R CMD check returned a NOTE that I addressed in this [PR](https://github.com/daranzolin/rperseus/pull/4).

- I always find useful to have a link to the URL for the API in the README, and
  potentially its documentation.
- How often is the catalog updated? Is it a concern that the package could get
  out of date and not include works that are being added to the catalog? Would
  it make sense to provide a function to generate the catalog locally?
- It would be good to validate that the user only provides a single `text_urn`
  to `get_perseus_text()` to provide an informative error message.
- The vignette indicates how to use `map_df` to fetch multiple texts at once,
  maybe add this information also in the README so it's obvious to users how to
  achieve this easily.
- What is the difference between the "Perseus Collection" and the "Perseus
  Catalog"? Why are some texts listed in the collection do not appear in the
  catalog? For instance, it seems that all the texts from the 19th century
  collection are not in the catalog. It might not be relevant for the package
  _per se_ but it might be informative to the users who might check the web
  interface first to find a work, only to realize they actually can't get it
  from the API.
- Running the checks of the package on r-hub, I see:
  * `NOTE: checking data for non-ASCII characters
     Note: found 61 marked UTF-8 strings`  (both on Windows and on R-devel on Linux)
  * `Error: processing vignette 'rperseus-vignette.Rmd' failed with diagnostics:
   could not find function "xpath_class"
   Execution halted` (on R-devel on Linux)
 (but I haven't had a chance to investigate the origin of these).
