### 2017-09-30 45d3fb9
updated README, added check
