### 2017-09-30 #145 @daranzolin
Thanks @fmichonneau! Comments below.

> I always find useful to have a link to the URL for the API in the README, and
potentially its documentation.

Done.

> How often is the catalog updated? Is it a concern that the package could get
out of date and not include works that are being added to the catalog? Would
it make sense to provide a function to generate the catalog locally?

I've contacted Tufts for clarification. In an earlier version of the package, `get_perseus_catalog` would generate the catalog locally, but I thought it would simplify things to just include the catalog on load.

> It would be good to validate that the user only provides a single text_urn
to get_perseus_text() to provide an informative error message.

Done.

> The vignette indicates how to use map_df to fetch multiple texts at once,
maybe add this information also in the README so it's obvious to users how to
achieve this easily.

I have an example of this under the tidyverse and tidytext heading.

> What is the difference between the "Perseus Collection" and the "Perseus
Catalog"? Why are some texts listed in the collection do not appear in the
catalog? For instance, it seems that all the texts from the 19th century
collection are not in the catalog. It might not be relevant for the package
per se but it might be informative to the users who might check the web
interface first to find a work, only to realize they actually can't get it
from the API.

The "Perseus Catalog" is admittedly my own moniker. And [the catalog XML I parsed for this package](http://www.perseus.tufts.edu/hopper/CTS?request=GetCapabilities) only included the Greek and Roman texts, but perhaps I should look further into incorporating the other literature. If the reviewers consider it worthwhile, I'll work on it next week.

 As for the r-hub checks, I will need further guidance on how to address them.
