### 2017-09-30 3082889
updated README for contributors
