+++ NAMESPACE
@@ -2,3 +2,4 @@
 
 export(get_perseus_text)
 importFrom(dplyr,"%>%")
+importFrom(rlang,.data)
