### 2017-10-15 #145 @fmichonneau
Hi @daranzolin, sorry for taking so long to get back to you. Thank you for addressing my comments.

It looks like you missed the comments I included under the checkboxes that I didn't check. Please address them, and I'll look more into the r-hub check results after that.
