### 2017-10-16 #145 @noamross
Let me just jump in to make sure we're on the same page.  @daranzolin, do you think at this point you've addressed everything from the @czeildi's and @fmichonneau's initial reviews? Let us know.  If so, @fmichonneau and @czeildi, let us know if you agree with a comment below and checking off remaining boxes on your reviews.  If not, let us know what else if left from your perspective.
