### 2017-10-17 #145 @daranzolin
Hi everyone, @noamross @czeildi @fmichonneau

Apologies for the delayed response--I see now that I have not yet addressed all of @fmichonneau 's initial review. I'll have responses/updates later this week.

Thanks for checking in!
