### 2017-10-23 fef032e
updated README, removed redundant docs
