+++ R/get_perseus_text.R
@@ -17,11 +17,8 @@
 #' @export
 #'
 #' @examples
-#' \dontrun{
-#' get_perseus_text("urn:cts:greekLit:tlg0013.tlg028.perseus-eng2")
 #' get_perseus_text("urn:cts:greekLit:tlg0013.tlg028.perseus-grc2")
 #' get_perseus_text("urn:cts:latinLit:stoa0215b.stoa003.opp-lat1")
-#' }
 
 get_perseus_text <- function(text_urn) {
 
