+++ man/perseus_catalog.Rd
@@ -1,8 +1,7 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R, R/rperseus.R
+% Please edit documentation in R/data.R
 \docType{data}
 \name{perseus_catalog}
-\alias{perseus_catalog}
 \alias{perseus_catalog}
 \title{Metadata for texts available via the Perseus Digital Library.}
 \format{A data frame with 2291 rows and 5 variables:
@@ -21,7 +20,5 @@ perseus_catalog
 }
 \description{
 A dataset containing the texts available from the Perseus Digital Library.
-
-Perseus Catalog
 }
 \keyword{datasets}
