+++ README.md
@@ -13,10 +13,15 @@ Author: David Ranzolin
 
 License: MIT
 
+Goal
+----
+
+The goal of `rperseus` is to furnish classicists, textual critics, and R enthusiasts with texts from the Classical World. While the English translations of most texts are available through `gutenbergr`, `rperseus`returns these works in their original language--Greek and Latin.
+
 Description
 -----------
 
-`rperseus` taps into the API endpoints at the [Perseus Digital Library's](http://www.perseus.tufts.edu/hopper/) CapiTainS environment. A wealth of primary texts and translations are available, from Homer to Cicero to Boetheius. The base API url is `http://cts.perseids.org/api/cts`.
+`rperseus` provides access to classical texts within the [Perseus Digital Library's](http://www.perseus.tufts.edu/hopper/) CapiTainS environment. A wealth of Greek and Latin texts are available, from Homer to Cicero to Boetheius. The Perseus Digital Library includes English translations in some cases. The base API url is `http://cts.perseids.org/api/cts`.
 
 Installation
 ------------
@@ -75,6 +80,7 @@ And here's how to retrieve the Greek text from Sophocles' underrated *Philoctete
 
 ``` r
 library(tidytext)
+#> Warning: package 'tidytext' was built under R version 3.4.2
 
 philoctetes <- perseus_catalog %>% 
   filter(group_name == "Sophocles",
