### 2017-10-23 #145 @daranzolin
Hi everyone, @noamross @czeildi @fmichonneau

Okay, I think I caught up where we left off:

> A statement of need clearly stating problems the software is designed to solve and its target audience in README

Updated README with a new **goal:** The goal of rperseus is to furnish classicists, textual critics, and R enthusiasts with texts from the Classical World. While the English translations of most texts are available through gutenbergr, rperseusreturns these works in their original language--Greek and Latin.

> in the documentation of the package, maybe start by describing the goal of
the package before writing about what it contains. You could also consider
pointing to the vignette to let users know how to get started.

I now link to the vignette in the README.

> I think the dataset (perseus_catalog) is documented twice, once in
R/data.R and once in R/rperseus.R.

Removed from `R/rperseus.R` Now only documented once.

I also heard back from Tufts regarding the updating of the catalog. The Managing Editor informed me that there is no regular update schedule at this time.
