### 2017-10-26 #145 @czeildi
Hi @daranzolin and @noamross ,

@daranzolin responded to my comments, I recommend to approve his package.

I have one minor comment: although the last chunk of the vignette is set to eval = FALSE it would be nice to change `map_df` to `purrr::map_df` or load purrr so that it can be run.

Great work!
