+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: rperseus
-Title: Get texts from the Perseus Digital Library
+Title: Get Texts from the Perseus Digital Library
 Version: 0.0.0.9000
 Authors@R: person("David", "Ranzolin", email = "daranzolin@gmail.com", role = c("aut", "cre"))
 Description: The Perseus Digital Library is a collection of classical texts. This package helps
