### 2017-10-26 1395b18
add links and use `code` in help files
