### 2017-10-26 #145 @noamross
One quick editor's comment, regarding the lazy-loaded catalog: There's sometimes reason to load stuff with the package for workflow reasons, but it's pretty rare because adding to the namespace is unexpected behavior for loading. I don't think this case merits going beyond the conventional ways of making data available: Via either `data("perseus_catalog")` or `perseus_catalog()` (the latter can simply take no arguments and return the data frame, or if you want to get fancy take some kind of filter or search as an argument).
