### 2017-10-26 93ce6e4
updated README with meta
