### 2017-10-26 ee8da93
commented out roxygen for get_full_text_index
