+++ man/rperseus.Rd
@@ -6,8 +6,8 @@
 \alias{rperseus-package}
 \title{rperseus: A package to obtain texts from the Perseus Digital Library}
 \description{
-The package contains a catalog and a function. The catalog is \code{perseus_catalog} and can be
-perused to locate a particular text and its corresponding URN. Then pass the URN to \code{get_perseus_text}
+The package contains a catalog and a function. The catalog is \code{\link{perseus_catalog}} and can be
+perused to locate a particular text and its corresponding URN. Then pass the URN to \code{\link{get_perseus_text}}
 to obtain the text.
 }
 \author{
