+++ man/get_full_text_index.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
-\name{get_full_text_index}
-\alias{get_full_text_index}
-\title{Gets the full index of a text (start to finish)}
-\usage{
-get_full_text_index(new_urn)
-}
-\arguments{
-\item{new_urn}{A reformatted urn}
-}
-\value{
-a character index (e.g. "1-110", "1.1-4.5")
-}
-\description{
-The API is flexible--you can append a text index to the end of your GET request
-of varying length. Rather than make multiple calls and iterate through smaller chunks of texts,
-why not get the entire work in one call?
-}
-\examples{
-\dontrun{
-get_full_text_index("latinLit/stoa0215b/stoa003/opp-lat1")
-}
-}
