### 2017-10-26 #145 @daranzolin
Thanks everyone! I'm giddy with excitement.

@czeildi -- added `library(purrr)` to README chunk.

@fmichonneau -- I merged pull request # 5 "Small fixes", commented out the roxygen for `get_full_text_index`, and added a Meta section to the README.

@noamross  -- Hadley recommends setting `LazyData: true` in [R Packages](http://r-pkgs.had.co.nz/data.html#data-data). Is this an instance where it should be false? Happy to make the change, but can you point me to an example of how to make the catalog available via `data` or `perseus_catalog`?
