### 2017-10-29 #145 @noamross
Great.  Sorry, @daranzolin, my brain was pulling up advice from old standards, you're right.  I've been typing `data('something')` I forget stuff is lazy-loaded anyway.

@fmichonneau, please check off the "responded and approved" box on your original review if everything looks on the up-and-up.
