### 2017-10-30 #145 @noamross
Approved! Thanks @daranzolin or submitting and @fmichonneau  and @czeildi for your reviews!

To-dos:
```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
```
- [ ] Add the "Peer-reviewed" status badge to your README:
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). _I_ for one am very interested in the types of analyses this package facilitates. Let us know if you are interested, and if so @stefaniebutland will be in touch about it.
