### 2017-10-31 03fd453
added ropensci footer
