+++ README.Rmd
@@ -18,6 +18,7 @@ knitr::opts_chunk$set(
 
 [![Build Status](https://travis-ci.org/daranzolin/rperseus.svg?branch=master)](https://travis-ci.org/daranzolin/rperseus)
 [![codecov](https://codecov.io/gh/daranzolin/rperseus/branch/master/graph/badge.svg)](https://codecov.io/gh/daranzolin/rperseus)
+[![](https://badges.ropensci.org/145_status.svg)](https://github.com/ropensci/onboarding/issues/145)
 
 
 ![](http://www.infobiblio.es/wp-content/uploads/2015/06/perseus-logo.png)
@@ -89,7 +90,7 @@ plato <- perseus_catalog %>%
 
 And here's how to retrieve the Greek text from Sophocles' underrated *Philoctetes* before unleashing the `tidytext` toolkit:
 
-```{r}
+```{r warning = FALSE}
 library(tidytext)
 
 philoctetes <- perseus_catalog %>% 
@@ -111,4 +112,7 @@ While there's no obvious way to filter out the Greek stop words and prepositions
 * [Report bugs or issues here.](https://github.com/daranzolin/rperseus/issues)
 * If you'd like to contribute to the development of `rperseus`, first get acquainted with the Perseus Digital Library, fork the repo, and send a pull request.
 * This project is released with a [Contributor Code of Conduct.](https://github.com/daranzolin/rperseus/blob/master/CONDUCT.md) By participating in this project, you agree to abide by its terms.
+
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
+
 
