### 2017-10-31 e0af890
updated README coverage links, updated pkgdown site
