### 2017-10-31 ece3f97
updated devtools::install_github()
