### 2017-10-31 #145 @daranzolin
Thanks @noamross ! This has been a wonderful, illuminating process, and I'm truly humbled by the attention of your reviewers.

I updated everything but have one lingering concern: [Travis isn't rebuilding from the latest commits.](https://travis-ci.org/ropensci/rperseus/builds) Do you know why that is?
