### 2017-11-01 ca740fb
updated DESCRIPTION links
