+++ DESCRIPTION
@@ -25,5 +25,5 @@ Suggests: knitr,
     roxygen2,
     tidytext
 VignetteBuilder: knitr
-URL: https://github.com/daranzolin/rperseus
-BugReports: https://github.com/daranzolin/rperseus/issues
+URL: https://github.com/ropesci/rperseus
+BugReports: https://github.com/ropensci/rperseus/issues
