### 2017-11-01 bac6e9d
corrected misspelled links
