### 2017-11-02 #145 @stefaniebutland
Hi @daranzolin. Are you leaning toward writing a blog post about `rperseus`?
>We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). I for one am very interested in the types of analyses this package facilitates.

I'm happy to answer any questions to help you decide
