### 2017-11-02 cafdea0
updated language variable in perseus_catalog
