### 2017-11-03 0fde7b2
implemented excerpt argument for get_perseus_text
