### 2017-11-03 d889ccd
updated README with excerpt example
