### 2017-11-03 #145 @daranzolin
Hi @stefaniebutland!

Yes, I am willing to do a blog. Disclaimer: I am not a scientist, and my use cases for this package are for literary analysis.

I am also attempting to add several new features, such as a morphological parser and a stop words dictionary. Perhaps postponing the blog may make it more interesting down the road?

Let me know what you think.
