### 2017-11-03 #145 @stefaniebutland
Great to hear that @daranzolin!

My strong bias would be to put something out sooner rather than later. The end part of the post could briefly outline your specific plans for new features (and maybe even point to a couple of issues that might attract new contributors?).

This is not a problem:
> Disclaimer: I am not a scientist, and my use cases for this package are for literary analysis.

I don't feel fully qualified to point out blog examples that directly relate to your niche but these might be relevant: https://ropensci.org/tags/text-analysis/ and several package / blog post contributors like Thomas Leeper and Lincoln Mullen are not "scientists" either.

Do you think a post in late November or early December would work for you?
Technical details: https://github.com/ropensci/roweb2#contributing-a-blog-post
