### 2017-11-03 #145 @daranzolin
@stefaniebutland late November is doable. I'll contact you if I encounter any roadblocks along the way.
