### 2017-11-11 090959c
included perseus_parallel, ggplot2 import
