+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(get_perseus_text)
+export(perseus_parallel)
+import(ggplot2)
 importFrom(dplyr,"%>%")
 importFrom(rlang,.data)
