### 2017-11-11 8bf351f
updated vignette
