### 2017-11-11 60aeb9a
included theme_perseus, split_every, updated internal catalog, fixed get_perseus_text
