### 2017-11-11 6530612
updated tests
