+++ man/get_perseus_text.Rd
@@ -4,10 +4,10 @@
 \alias{get_perseus_text}
 \title{Get a primary text by URN.}
 \usage{
-get_perseus_text(text_urn, excerpt = NULL)
+get_perseus_text(urn, excerpt = NULL)
 }
 \arguments{
-\item{text_urn}{Valid uniform resource number (URN) obtained from \code{\link{perseus_catalog}}.}
+\item{urn}{Valid uniform resource number (URN) obtained from \code{\link{perseus_catalog}}.}
 
 \item{excerpt}{An index to excerpt the text. For example, the first four "verses" of a text might be 1.1-1.4. If NULL, the entire work is returned.}
 }
@@ -30,5 +30,5 @@ Get a primary text by URN.
 \examples{
 get_perseus_text("urn:cts:greekLit:tlg0013.tlg028.perseus-grc2")
 get_perseus_text("urn:cts:latinLit:stoa0215b.stoa003.opp-lat1")
-get_perseus_text(text_urn = "urn:cts:greekLit:tlg0031.tlg009.perseus-grc2", excerpt = "5.1-5.5")
+get_perseus_text(urn = "urn:cts:greekLit:tlg0031.tlg009.perseus-grc2", excerpt = "5.1-5.5")
 }
