### 2017-11-26 f2840af
implemented parse_excerpt
