+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(get_perseus_text)
+export(parse_excerpt)
 export(perseus_parallel)
 import(ggplot2)
 importFrom(dplyr,"%>%")
