### 2017-11-27 7821c3b
updated README, vignette with parse_excerpt example
