+++ README.md
@@ -76,6 +76,225 @@ qoheleth$text
 #> [1] "דִּבְרֵי֙ קֹהֶ֣לֶת בֶּן־ דָּוִ֔ד מֶ֖לֶךְ בִּירוּשָׁלִָֽם : הֲבֵ֤ל הֲבָלִים֙ אָמַ֣ר קֹהֶ֔לֶת הֲבֵ֥ל הֲבָלִ֖ים הַכֹּ֥ל הָֽבֶל : מַה־ יִּתְר֖וֹן לָֽאָדָ֑ם בְּכָל־ עֲמָל֔וֹ שֶֽׁיַּעֲמֹ֖ל תַּ֥חַת הַשָּֽׁמֶשׁ :"
 ```
 
+Parsing Excerpts
+----------------
+
+You can parse any Greek excerpt, returning a data frame with each word's part of speech, gender, case, mood, voice, tense, person, number, and degree.
+
+``` r
+knitr::kable(parse_excerpt("urn:cts:greekLit:tlg0031.tlg002.perseus-grc2", "5.1-5.2"))
+```
+
+| word         | form           | verse | part\_of\_speech | person | number   | tense     | mood        | voice         | gender    | case       | degree |
+|:-------------|:---------------|:------|:-----------------|:-------|:---------|:----------|:------------|:--------------|:----------|:-----------|:-------|
+| καί          | Καὶ            | 5.1   | conjunction      | NA     | NA       | NA        | NA          | NA            | NA        | NA         | NA     |
+| ἔρχομαι      | ἦλθον          | 5.1   | verb             | third  | plural   | aorist    | indicative  | active        | NA        | NA         | NA     |
+| εἰς          | εἰς            | 5.1   | preposition      | NA     | NA       | NA        | NA          | NA            | NA        | NA         | NA     |
+| ὁ            | τὸ             | 5.1   | article          | NA     | singular | NA        | NA          | NA            | neuter    | accusative | NA     |
+| πέραν        | πέραν          | 5.1   | adverb           | NA     | NA       | NA        | NA          | NA            | NA        | NA         | NA     |
+| ὁ            | τῆς            | 5.1   | article          | NA     | singular | NA        | NA          | NA            | feminine  | genative   | NA     |
+| θάλασσα      | θαλάσσης       | 5.1   | noun             | NA     | singular | NA        | NA          | NA            | feminine  | genative   | NA     |
+| εἰς          | εἰς            | 5.1   | preposition      | NA     | NA       | NA        | NA          | NA            | NA        | NA         | NA     |
+| ὁ            | τὴν            | 5.1   | article          | NA     | singular | NA        | NA          | NA            | feminine  | accusative | NA     |
+| χώρα         | χώραν          | 5.1   | noun             | NA     | singular | NA        | NA          | NA            | feminine  | accusative | NA     |
+| ὁ            | τῶν            | 5.1   | article          | NA     | plural   | NA        | NA          | NA            | masculine | genative   | NA     |
+| Γερασηνός    | Γερασηνῶν      | 5.1   | noun             | NA     | plural   | NA        | NA          | NA            | masculine | genative   | NA     |
+| .            | .              | 5.1   | punctuation      | NA     | NA       | NA        | NA          | NA            | NA        | NA         | NA     |
+| καί          | καὶ            | 5.2   | adverb           | NA     | NA       | NA        | NA          | NA            | NA        | NA         | NA     |
+| ἐξέρχομαι    | ἐξελθόντος     | 5.2   | verb             | NA     | singular | aorist    | participle  | active        | masculine | genative   | NA     |
[…]
