### 2017-11-27 bb5582d
modified filter_list to sequence correctly
