### 2017-11-27 02d0ed4
updated version, pkgdown site, README, NEWS
