+++ NEWS.md
@@ -1,7 +1,15 @@
+# rperseus 0.1.2
+
+* Transferred repo ownership to rOpenSci
+* Includes `perseus_parallel`, a function to render parallels with `ggplot2`
+* Includes `parse_excerpt`, a function to parse any Greek excerpt
+* Updated README, vignette
+
 # rperseus 0.1
 
-* First version of package, including
+First version of package, including
   * `get_perseus_text` function, for getting texts from the Perseus Digital Library
   * `perseus_catalog` data frame to peruse text URNs
   * Introductory vignette including basic examples of retrieving texts
-* Added a `NEWS.md` file to track changes to the package.
+  * Added a `NEWS.md` file to track changes to the package
+  * `pkgdown` site
