+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: rperseus
 Title: Get Texts from the Perseus Digital Library
-Version: 0.0.0.9000
+Version: 0.1.2
 Authors@R: person("David", "Ranzolin", email = "daranzolin@gmail.com", role = c("aut", "cre"))
 Description: The Perseus Digital Library is a collection of classical texts. This package helps
     you get them. The available works can also be viewed here: <http://cts.perseids.org/>.
