+++ man/greek_stop_words.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{greek_stop_words}
+\alias{greek_stop_words}
+\title{A dictionary of Greek stop words}
+\format{A data frame with 223 rows and 1 variable:
+\describe{
+  \item{word}{Greek stop word}
+}}
+\source{
+\url{Compiled manually by filtering prepositions, pronouns, conjunctions, particles, and articles.}
+}
+\usage{
+greek_stop_words
+}
+\description{
+A dictionary of Greek stop words
+}
+\keyword{datasets}
