### 2017-11-28 463bc05
updated pkgdown site, README
