### 2017-11-28 9d9b5b4
updated vignette, data docs
